+++ crates/modelardb_server/src/error.rs
@@ -27,7 +27,6 @@ use datafusion::error::DataFusionError;
 use deltalake::errors::DeltaTableError;
 use modelardb_common::error::ModelarDbCommonError;
 use modelardb_storage::error::ModelarDbStorageError;
-use modelardb_types::error::ModelarDbTypesError;
 use object_store::Error as ObjectStoreError;
 use tonic::transport::Error as TonicTransportError;
 use tonic::Status as TonicStatusError;
@@ -60,8 +59,6 @@ pub enum ModelarDbServerError {
     ModelarDbCommon(ModelarDbCommonError),
     /// Error returned by modelardb_storage.
     ModelarDbStorage(ModelarDbStorageError),
-    /// Error returned by modelardb_types.
-    ModelarDbTypes(ModelarDbTypesError),
     /// Status returned by Tonic.
     TonicStatus(TonicStatusError),
     /// Error returned by Tonic.
@@ -82,7 +79,6 @@ impl Display for ModelarDbServerError {
             Self::ObjectStore(reason) => write!(f, "Object Store Error: {reason}"),
             Self::ModelarDbCommon(reason) => write!(f, "ModelarDB Common Error: {reason}"),
             Self::ModelarDbStorage(reason) => write!(f, "ModelarDB Storage Error: {reason}"),
-            Self::ModelarDbTypes(reason) => write!(f, "ModelarDB Types Error: {reason}"),
             Self::TonicStatus(reason) => write!(f, "Tonic Status Error: {reason}"),
             Self::TonicTransport(reason) => write!(f, "Tonic Transport Error: {reason}"),
         }
@@ -104,7 +100,6 @@ impl Error for ModelarDbServerError {
             Self::ObjectStore(reason) => Some(reason),
             Self::ModelarDbCommon(reason) => Some(reason),
             Self::ModelarDbStorage(reason) => Some(reason),
-            Self::ModelarDbTypes(reason) => Some(reason),
             Self::TonicStatus(reason) => Some(reason),
             Self::TonicTransport(reason) => Some(reason),
         }
@@ -156,12 +151,6 @@ impl From<ModelarDbCommonError> for ModelarDbServerError {
 impl From<ModelarDbStorageError> for ModelarDbServerError {
     fn from(error: ModelarDbStorageError) -> Self {
         Self::ModelarDbStorage(error)
-    }
-}
-
-impl From<ModelarDbTypesError> for ModelarDbServerError {
-    fn from(error: ModelarDbTypesError) -> Self {
-        Self::ModelarDbTypes(error)
     }
 }
